Force cURL to use IPv4. Avoid IPv6 server name
If the server is configured to make outgoing requests with IPv6 whenever possible, then the `curl` command to `canhazip.com` will return the server's IPv6 URL.

This is undesirable, because this script only configures OpenVPN for IPv4. A more long-term fix would be to enable OpenVPN over IPv6, in the face of IPv4 address exhaustion.